requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
dj-database-url==1.3.0
Django==4.2
django-filter==23.1
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.4
joblib==1.2.0
Markdown==3.4.3
numpy==1.24.3
psycopg2==2.9.6
python-dateutil==1.5
pytz==2023.3
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
sqlparse==0.4.4
threadpoolctl==3.1.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
whitenoise==6.4.0
